Moved setting default source for tasks to tfdyn
