+++ acceptance/bundle/deploy/jobs/task-source/databricks.yml
@@ -1,19 +1,21 @@
 bundle:
   name: task-source
 
+git_source: &git_source
+  git_source:
+    git_branch: main
+    git_url: https://github.com/databricks/cli.git
+    git_provider: github
+
 resources:
   jobs:
     git_job:
       name: git_job
-      git_source:
-        git_branch: main
-        git_url: https://github.com/databricks/cli.git
-        git_provider: github
-
+      <<: *git_source
       tasks:
         - task_key: test_task_no_source
           notebook_task:
-            notebook_path: ./notebook
+            notebook_path: ./notebook.py
         - task_key: test_task_source_workspace
           notebook_task:
             notebook_path: ./notebook.py
