+++ acceptance/bundle/deploy/jobs/task-source/output.txt
@@ -2,7 +2,7 @@
 >>> [CLI] bundle deploy
 Warning: invalid value "github" for enum field. Valid values are [awsCodeCommit azureDevOpsServices bitbucketCloud bitbucketServer gitHub gitHubEnterprise gitLab gitLabEnterpriseEdition]
   at resources.jobs.git_job.git_source.git_provider
-  in databricks.yml:11:23
+  in databricks.yml:8:19
 
 Uploading bundle files to /Workspace/Users/[USERNAME]/.bundle/task-source/default/files...
 Deploying resources...
@@ -10,100 +10,20 @@ Updating deployment state...
 Deployment complete!
 
 >>> jq -s .[] | select(.path=="/api/2.2/jobs/create") | select(.body.name=="git_job") | .body out.requests.txt
-{
-  "deployment": {
-    "kind": "BUNDLE",
-    "metadata_file_path": "/Workspace/Users/[USERNAME]/.bundle/task-source/default/state/metadata.json"
-  },
-  "edit_mode": "UI_LOCKED",
-  "format": "MULTI_TASK",
-  "git_source": {
-    "git_branch": "main",
-    "git_provider": "github",
-    "git_url": "https://github.com/databricks/cli.git"
-  },
-  "max_concurrent_runs": 1,
-  "name": "git_job",
-  "queue": {
-    "enabled": true
-  },
-  "tasks": [
-    {
-      "for_each_task": {
-        "inputs": "[1]",
-        "task": {
-          "notebook_task": {
-            "notebook_path": "./notebook.py",
-            "source": "GIT"
-          },
-          "task_key": "test_task_foreach_no_source"
-        }
-      },
-      "task_key": "test_task_foreach_no_source"
-    },
-    {
-      "for_each_task": {
-        "inputs": "[1]",
-        "task": {
-          "notebook_task": {
-            "notebook_path": "./notebook.py",
-            "source": "WORKSPACE"
-          },
-          "task_key": "test_task_foreach_workspace"
-        }
-      },
-      "task_key": "test_task_foreach_workspace"
-    },
-    {
-      "notebook_task": {
-        "notebook_path": "./notebook",
-        "source": "GIT"
-      },
-      "task_key": "test_task_no_source"
-    },
-    {
-      "notebook_task": {
-        "notebook_path": "./notebook.py",
-        "source": "WORKSPACE"
-      },
-      "task_key": "test_task_source_workspace"
-    }
-  ]
-}
 
 >>> jq -s .[] | select(.path=="/api/2.2/jobs/create") | select(.body.name=="workspace_job") | .body out.requests.txt
-{
-  "deployment": {
-    "kind": "BUNDLE",
-    "metadata_file_path": "/Workspace/Users/[USERNAME]/.bundle/task-source/default/state/metadata.json"
-  },
-  "edit_mode": "UI_LOCKED",
-  "format": "MULTI_TASK",
-  "max_concurrent_runs": 1,
-  "name": "workspace_job",
-  "queue": {
-    "enabled": true
-  },
-  "tasks": [
-    {
-      "for_each_task": {
-        "inputs": "[1]",
-        "task": {
-          "notebook_task": {
-            "notebook_path": "/Workspace/Users/[USERNAME]/.bundle/task-source/default/files/notebook",
-            "source": "WORKSPACE"
-          },
-          "task_key": "test_task_foreach_no_source"
-        }
-      },
-      "task_key": "test_task_foreach"
-    },
-    {
-      "notebook_task": {
-        "notebook_path": "/Workspace/Users/[USERNAME]/.bundle/task-source/default/files/notebook",
-        "source": "WORKSPACE"
-      },
-      "task_key": "test_task_no_source"
-    }
-  ]
-}
+
+>>> update_file.py databricks.yml <<: *git_source 
+
+>>> [CLI] bundle plan
+update jobs.git_job
+
+>>> [CLI] bundle deploy
+Uploading bundle files to /Workspace/Users/[USERNAME]/.bundle/task-source/default/files...
+Deploying resources...
+Updating deployment state...
+Deployment complete!
+
+>>> jq -s .[] | select(.path=="/api/2.2/jobs/reset") | select(.body.new_settings.name=="git_job") | .body out.requests.txt
+
+>>> jq -s .[] | select(.path=="/api/2.2/jobs/reset") | select(.body.new_settings.name=="workspace_job") | .body out.requests.txt
