+++ acceptance/bundle/deploy/jobs/task-source/test.toml
@@ -2,3 +2,4 @@ Cloud = false
 Local = true
 
 RecordRequests = true
+EnvVaryOutput = "DATABRICKS_CLI_DEPLOYMENT"
