+++ bundle/config/mutator/resourcemutator/resource_mutator.go
@@ -115,8 +115,6 @@ func applyInitializeMutators(ctx context.Context, b *bundle.Bundle) {
 		// Updates (dynamic): resources.*.*.permissions (removes permissions entries where user_name or service_principal_name matches current user)
 		// Removes the current user from all resource permissions as the Terraform provider implicitly grants ownership
 		FilterCurrentUser(),
-
-		ApplyDefaultTaskSource(),
 	)
 }
